requirements.txt
boto3
mysql_connector
retrying
